+++ app/src/main/java/com/example/vknewsclient/ui/theme/HomeScreen.kt
@@ -1,0 +1,66 @@
+package com.example.vknewsclient.ui.theme
+
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.DismissDirection
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.SwipeToDismiss
+import androidx.compose.material3.rememberDismissState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.example.vknewsclient.MainViewModel
+
+@OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
+@Composable
+fun HomeScreen(
+    viewModel: MainViewModel
+) {
+    val feedPosts = viewModel.feedPost.observeAsState(listOf())
+
+    LazyColumn(
+        contentPadding = PaddingValues(
+            top = 16.dp,
+            start = 8.dp,
+            end = 8.dp,
+            bottom = 92.dp
+        ),
+        verticalArrangement = Arrangement.spacedBy(8.dp)
+    ) {
+        items(items = feedPosts.value, key = { it.id }) {feedPost ->
+            val dissmissState = rememberDismissState()
+
+            if (dissmissState.isDismissed(direction = DismissDirection.EndToStart)) {
+                viewModel.removePost(feedPost)
+            }
+
+            SwipeToDismiss(
+                modifier = Modifier.animateItemPlacement(),
+                state = dissmissState,
+                background = {},
+                directions = setOf(DismissDirection.EndToStart),
+                dismissContent = {
+                    PostCard(
+                        feedPost = feedPost,
+                        onViewsClickListener = { statisticItem ->
+                            viewModel.updateCount(feedPost, statisticItem)
+                        },
+                        onLikeClickListener = { statisticItem ->
+                            viewModel.updateCount(feedPost, statisticItem)
+                        },
+                        onCommentClickListener = { statisticItem ->
+                            viewModel.updateCount(feedPost, statisticItem)
+                        },
+                        onShareClickListener = { statisticItem ->
+                            viewModel.updateCount(feedPost, statisticItem)
+                        }
+                    )
+                }
+            )
+        }
+    }
+}
