+++ app/src/main/java/com/example/vknewsclient/ui/theme/Theme.kt
@@ -55,7 +55,7 @@ fun VkNewsClientTheme(
         SideEffect {
             val window = (view.context as Activity).window
             window.statusBarColor = colorScheme.primary.toArgb()
-            WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = darkTheme
+            WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = !darkTheme
         }
     }
