+++ app/src/main/java/com/example/vknewsclient/ui/theme/VKNewsMainScreen.kt
@@ -1,50 +1,37 @@
 package com.example.vknewsclient.ui.theme
 
 import android.annotation.SuppressLint
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.DismissDirection
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.NavigationBarItemDefaults
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SwipeToDismiss
 import androidx.compose.material3.Text
-import androidx.compose.material3.rememberDismissState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import com.example.vknewsclient.MainViewModel
 
 @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
-@OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
 @Composable
 fun MainScreen(viewModel: MainViewModel) {
+    val selectedNavItem by viewModel.selectedNavItem.observeAsState(NavigationItem.Home)
+
     Scaffold(
         bottomBar = {
             NavigationBar {
-                val selectedItemPosition = rememberSaveable {
-                    mutableIntStateOf(0)
-                }
                 val items = listOf(
                     NavigationItem.Home,
                     NavigationItem.Favourite,
                     NavigationItem.Profile
                 )
-                items.forEachIndexed { index, item ->
+                items.forEach { item ->
                     NavigationBarItem(
-                        selected = index == selectedItemPosition.intValue,
-                        onClick = { selectedItemPosition.intValue = index },
+                        selected = selectedNavItem == item,
+                        onClick = { viewModel.selectNavItem(item) },
                         icon = {
                             Icon(item.icon, contentDescription = null)
                         },
@@ -63,48 +50,10 @@ fun MainScreen(viewModel: MainViewModel) {
             }
         },
         content = {
-            val feedPosts = viewModel.feedPost.observeAsState(listOf())
-
-            LazyColumn(
-                contentPadding = PaddingValues(
-                    top = 16.dp,
-                    start = 8.dp,
-                    end = 8.dp,
-                    bottom = 92.dp
-                ),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                items(items = feedPosts.value, key = { it.id }) {feedPost ->
-                    val dissmissState = rememberDismissState()
-
-                    if (dissmissState.isDismissed(direction = DismissDirection.EndToStart)) {
-                        viewModel.removePost(feedPost)
-                    }
-
-                    SwipeToDismiss(
-                        modifier = Modifier.animateItemPlacement(),
-                        state = dissmissState,
-                        background = {},
-                        directions = setOf(DismissDirection.EndToStart),
-                        dismissContent = {
-                            PostCard(
-                                feedPost = feedPost,
-                                onViewsClickListener = { statisticItem ->
-                                    viewModel.updateCount(feedPost, statisticItem)
-                                },
-                                onLikeClickListener = { statisticItem ->
-                                    viewModel.updateCount(feedPost, statisticItem)
-                                },
-                                onCommentClickListener = { statisticItem ->
-                                    viewModel.updateCount(feedPost, statisticItem)
-                                },
-                                onShareClickListener = { statisticItem ->
-                                    viewModel.updateCount(feedPost, statisticItem)
-                                }
-                            )
-                        }
-                    )
-                }
+            when (selectedNavItem) {
+                NavigationItem.Home -> HomeScreen(viewModel = viewModel)
+                NavigationItem.Favourite -> Text(text = "Favourite", color = Color.Blue)
+                NavigationItem.Profile -> Text(text = "Profile", color = Color.Blue)
             }
         }
     )
